Rename s11n to internal/s11n
