+++ go.mod
@@ -1,11 +1,10 @@
-module githu.com/hashcloak/katzenmint-pki
+module github.com/hashcloak/katzenmint-pki
 
 go 1.15
 
 require (
 	github.com/dgraph-io/badger v1.6.1
 	github.com/google/go-cmp v0.5.4 // indirect
-	github.com/hashcloak/katzenmint-pki v0.0.0-20200825134709-8604e7226b71
 	// github.com/hashcloak/katzenmint-pki v0.0.0-20200825134709-8604e7226b71
 	github.com/katzenpost/core v0.0.11
 	github.com/kr/text v0.2.0 // indirect
