+++ internal/s11n/descriptor.go
